Separate MySQL from Mariadb implementations
