+++ rm-datasource/src/main/java/io/seata/rm/datasource/exec/ExecuteTemplate.java
@@ -25,6 +25,8 @@ import io.seata.common.util.CollectionUtils;
 import io.seata.core.context.RootContext;
 import io.seata.core.model.BranchType;
 import io.seata.rm.datasource.StatementProxy;
+import io.seata.rm.datasource.exec.mariadb.MariadbInsertOnDuplicateUpdateExecutor;
+import io.seata.rm.datasource.exec.mariadb.MariadbUpdateJoinExecutor;
 import io.seata.rm.datasource.exec.mysql.MySQLInsertOnDuplicateUpdateExecutor;
 import io.seata.rm.datasource.exec.mysql.MySQLUpdateJoinExecutor;
 import io.seata.rm.datasource.sql.SQLVisitorFactory;
@@ -107,9 +109,12 @@ public class ExecuteTemplate {
                     case INSERT_ON_DUPLICATE_UPDATE:
                         switch (dbType) {
                             case JdbcConstants.MYSQL:
-                            case JdbcConstants.MARIADB:
                                 executor =
                                     new MySQLInsertOnDuplicateUpdateExecutor(statementProxy, statementCallback, sqlRecognizer);
+                                break;
+                            case JdbcConstants.MARIADB:
+                                executor =
+                                    new MariadbInsertOnDuplicateUpdateExecutor(statementProxy, statementCallback, sqlRecognizer);
                                 break;
                             default:
                                 throw new NotSupportYetException(dbType + " not support to INSERT_ON_DUPLICATE_UPDATE");
@@ -119,6 +124,9 @@ public class ExecuteTemplate {
                         switch (dbType) {
                             case JdbcConstants.MYSQL:
                                 executor = new MySQLUpdateJoinExecutor<>(statementProxy,statementCallback,sqlRecognizer);
+                                break;
+                            case JdbcConstants.MARIADB:
+                                executor = new MariadbUpdateJoinExecutor<>(statementProxy,statementCallback,sqlRecognizer);
                                 break;
                             default:
                                 throw new NotSupportYetException(dbType + " not support to " + SQLType.UPDATE_JOIN.name());
