+++ rm-datasource/src/main/java/io/seata/rm/datasource/exec/mariadb/MariadbUpdateJoinExecutor.java
@@ -15,49 +15,17 @@
  */
 package io.seata.rm.datasource.exec.mariadb;
 
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.StringJoiner;
 
 import io.seata.rm.datasource.exec.mysql.MySQLUpdateJoinExecutor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import io.seata.common.exception.ShouldNeverHappenException;
-import io.seata.common.util.CollectionUtils;
-import io.seata.common.util.IOUtil;
-import io.seata.common.util.StringUtils;
-import io.seata.core.protocol.Version;
-import io.seata.rm.datasource.ConnectionProxy;
-import io.seata.rm.datasource.SqlGenerateUtils;
 import io.seata.rm.datasource.StatementProxy;
 import io.seata.rm.datasource.exec.StatementCallback;
-import io.seata.rm.datasource.exec.UpdateExecutor;
-import io.seata.rm.datasource.sql.struct.Field;
-import io.seata.rm.datasource.sql.struct.TableMeta;
-import io.seata.rm.datasource.sql.struct.TableMetaCacheFactory;
-import io.seata.rm.datasource.sql.struct.TableRecords;
-import io.seata.rm.datasource.undo.SQLUndoLog;
-import io.seata.sqlparser.JoinRecognizer;
-import io.seata.sqlparser.ParametersHolder;
 import io.seata.sqlparser.SQLRecognizer;
-import io.seata.sqlparser.SQLType;
-import io.seata.sqlparser.SQLUpdateRecognizer;
-import io.seata.sqlparser.util.ColumnUtils;
-
 
 /**
  * @author jianbin.chen
  */
 public class MariadbUpdateJoinExecutor<T, S extends Statement> extends MySQLUpdateJoinExecutor<T, S> {
-    private static final Logger LOGGER = LoggerFactory.getLogger(MariadbUpdateJoinExecutor.class);
 
     /**
      * Instantiates a new Update executor.
@@ -69,5 +37,7 @@ public class MariadbUpdateJoinExecutor<T, S extends Statement> extends MySQLUpda
     public MariadbUpdateJoinExecutor(StatementProxy<S> statementProxy, StatementCallback<T, S> statementCallback,
         SQLRecognizer sqlRecognizer) {
         super(statementProxy, statementCallback, sqlRecognizer);
+        this.isLowerSupportGroupByPksVersion = false;
     }
+
 }
