+++ rm-datasource/src/main/java/io/seata/rm/datasource/undo/mariadb/MySQLUndoLogManager.java
@@ -13,12 +13,15 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.rm.datasource.undo.mysql;
+package io.seata.rm.datasource.undo.mariadb;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.Date;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import io.seata.common.loader.LoadLevel;
 import io.seata.core.compressor.CompressorType;
@@ -26,8 +29,6 @@ import io.seata.core.constants.ClientTableColumnsName;
 import io.seata.rm.datasource.undo.AbstractUndoLogManager;
 import io.seata.rm.datasource.undo.UndoLogParser;
 import io.seata.sqlparser.util.JdbcConstants;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * @author jsbxyyx
