+++ rm-datasource/src/main/java/io/seata/rm/datasource/undo/mariadb/keyword/MySQLKeywordChecker.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.rm.datasource.undo.mysql.keyword;
+package io.seata.rm.datasource.undo.mariadb.keyword;
 
 import java.util.Arrays;
 import java.util.Set;
