+++ rm-datasource/src/main/resources/META-INF/services/io.seata.rm.datasource.sql.struct.TableMetaCache
@@ -1,3 +1,4 @@
 io.seata.rm.datasource.sql.struct.cache.MysqlTableMetaCache
+io.seata.rm.datasource.sql.struct.cache.MariadbTableMetaCache
 io.seata.rm.datasource.sql.struct.cache.OracleTableMetaCache
 io.seata.rm.datasource.sql.struct.cache.PostgresqlTableMetaCache
